Start goals and spec, add example and structure
Start working on the list of goals (and non-goals) for this spec.

Also start working on the spec, defining the document and the root
element.

Also included is a (small) section in the intro, explaining how the
document is laid out.  The first example is also added, showing a
document authoritatively saying "No endpoints serve this path."